Add FRED key config and central log DB
- include FRED_API_KEY in config files\n- move WallStreetBets analysis into scrapers and add CLI\n- database logger writes to new system_logs table
